server/src/test/java/learn/capybara_mapper/data/UserJdbcTemplateRepositoryTest.java
package learn.capybara_mapper.data;

import learn.capybara_mapper.models.User;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.BeforeEach;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
class UserJdbcTemplateRepositoryTest {

    @Autowired
    UserJdbcTemplateRepository repository;

    @Autowired
    KnownGoodState knownGoodState;

    @BeforeEach
    void setup() {
        knownGoodState.set();
    }
    @Test
    void shouldFindAll() {
        List<User> users = repository.findAll();
        assertNotNull(users);
        assertEquals(4, users.size());
    }

    @Test
    void shouldFindById() {
        User user = repository.findById(3);
        assertNotNull(user);
        assertEquals(3, user.getUserId());
        assertEquals("Ryann Jacobs", user.getName());
    }
}